Incorporated the suggested review comments.
